Begin breaking up backend
